+++ pkg/segment/writer/metrics/wal/wal_test.go
@@ -86,7 +86,7 @@ func generateRandomDatapoints(n int) []WalDatapoint {
 	currentMillis := time.Now().UnixMilli()
 	for i := 0; i < n; i++ {
 		dp := WalDatapoint{
-			Timestamp: uint64(currentMillis + int64(i*1000)),
+			Timestamp: uint32(currentMillis + int64(i*1000)),
 			DpVal:     float64(10 + i),
 			Tsid:      uint64(i + 1),
 		}
